Functions minimized and clean code

+++ l4_io/SSD1306_OLED/SSD1306_OLED_ascii/SSD1306_OLED_ascii.h
@@ -55,4 +55,4 @@ void SSD1306_ascii_display_Z();
 
 void SSD1306_ascii_display_space();
 
-void SSD1306_ascii_display_driver(uint8_t ascii_letter_size, uint16_t ascii_letter[]);
+void SSD1306_ascii_display_driver(uint8_t ascii_letter_size, uint8_t ascii_letter[]);
